pyproject.toml
[tool.poetry]
name = "matching_pairs"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
google-cloud-bigquery = "^3.29.0"
google-auth = "^2.38.0"
boto3 = "^1.36.10"
pandas = "^2.2.3"
db-dtypes = "^1.4.0"
gspread = "^6.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pysen = "^0.10.4"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py310"

[[tool.pysen.lint.mypy_targets]]
  paths = ["."]
